Mudanças de estilo de código.

+++ io-board/stm32f4/base/lib/provant/pv_typedefs.h
@@ -87,9 +87,9 @@ typedef struct
 /** \brief Tipo do Servo com informações. (Iuro)*/
 typedef struct
 {
-	uint32_t servo_id;
+	uint8_t servo_id;
 	uint32_t heartBeat;
-	int32_t  pwm;
+	int16_t  pwm;
 	float angularSpeed;
 	float position;
 } pv_msg_servo;
